requirements.txt
pytube
pytubefix~=8.12.3
requests~=2.32.3
pillow~=11.0.0